+++ Editor/Drawers/SaintsEventBaseTypeDrawer/PersistentArgumentDrawer.cs
@@ -1,3 +1,4 @@
+#if SAINTSFIELD_SERIALIZATION && !SAINTSFIELD_SERIALIZATION_DISABLED && UNITY_2022_2_OR_NEWER && !SAINTSFIELD_UI_TOOLKIT_DISABLE
 using SaintsField.Editor.Core;
 using SaintsField.Events;
 using UnityEditor;
@@ -13,3 +14,4 @@ namespace SaintsField.Editor.Drawers.SaintsEventBaseTypeDrawer
 
     }
 }
+#endif