+++ Runtime/Events/SaintsEventBase.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using UnityEngine;
+using UnityEngine.Events;
 
 // ReSharper disable once CheckNamespace
 namespace SaintsField.Events
@@ -29,6 +30,23 @@ namespace SaintsField.Events
         public void RemoveAllListeners()
         {
             RuntimeCalls.Clear();
+        }
+
+        public int GetPersistentEventCount() => _persistentCalls.Length;
+
+        public UnityEventCallState GetPersistentListenerState(int index) => _persistentCalls[index].callState;
+
+        public string GetPersistentMethodName(int index) => _persistentCalls[index].methodName;
+
+        public UnityEngine.Object GetPersistentTarget(int index)
+        {
+            PersistentCall call = _persistentCalls[index];
+            return call.isStatic ? null : call.target;
+        }
+
+        public void SetPersistentListenerState(int index, UnityEventCallState state)
+        {
+            _persistentCalls[index].callState = state;
         }
     }
 }